Split code to several files. Introduce strawman.collection.mutable and strawman.collection.immutable packages. Seq is still in strawman.collection package because it is implemented by both mutable (e.g. ArrayBuffer) and immutable (e.g. List) concrete collections.
