+++ src/main/scala/strawman/collection/IterableOnce.scala
@@ -1,0 +1,8 @@
+package strawman.collection
+
+import strawman.collection.mutable.Iterator
+
+trait IterableOnce[+A] {
+  /** Iterator can be used only once */
+  def iterator(): Iterator[A]
+}
