+++ src/main/scala/strawman/collection/mutable/Iterator.scala
@@ -1,0 +1,111 @@
+package strawman.collection.mutable
+
+import scala.{Boolean, Int, Unit, Nothing, NoSuchElementException}
+import strawman.collection.{IndexedView, IterableOnce}
+
+/** A core Iterator class */
+trait Iterator[+A] { self =>
+  def hasNext: Boolean
+  def next(): A
+  def foldLeft[B](z: B)(op: (B, A) => B): B =
+    if (hasNext) foldLeft(op(z, next()))(op) else z
+  def foldRight[B](z: B)(op: (A, B) => B): B =
+    if (hasNext) op(next(), foldRight(z)(op)) else z
+  def foreach(f: A => Unit): Unit =
+    while (hasNext) f(next())
+  def indexWhere(p: A => Boolean): Int = {
+    var i = 0
+    while (hasNext) {
+      if (p(next())) return i
+      i += 1
+    }
+    -1
+  }
+  def length = {
+    var len = 0
+    while (hasNext) { len += 1; next() }
+    len
+  }
+  def filter(p: A => Boolean): Iterator[A] = new Iterator[A] {
+    private var hd: A = _
+    private var hdDefined: Boolean = false
+
+    def hasNext: Boolean = hdDefined || {
+      do {
+        if (!self.hasNext) return false
+        hd = self.next()
+      } while (!p(hd))
+      hdDefined = true
+      true
+    }
+
+    def next() =
+      if (hasNext) {
+        hdDefined = false
+        hd
+      }
+      else Iterator.empty.next()
+  }
+  def map[B](f: A => B): Iterator[B] = new Iterator[B] {
+    def hasNext = self.hasNext
+    def next() = f(self.next())
+  }
+
+  def flatMap[B](f: A => IterableOnce[B]): Iterator[B] = new Iterator[B] {
+    private var myCurrent: Iterator[B] = Iterator.empty
+    private def current = {
+      while (!myCurrent.hasNext && self.hasNext)
+        myCurrent = f(self.next()).iterator()
+      myCurrent
+    }
+    def hasNext = current.hasNext
+    def next() = current.next()
+  }
+  def ++[B >: A](xs: IterableOnce[B]): Iterator[B] = new Iterator[B] {
+    private var myCurrent: Iterator[B] = self
+    private var first = true
+    private def current = {
+      if (!myCurrent.hasNext && first) {
+        myCurrent = xs.iterator()
+        first = false
+      }
+      myCurrent
+    }
+    def hasNext = current.hasNext
+    def next() = current.next()
+  }
+  def take(n: Int): Iterator[A] = new Iterator[A] {
+    private var i = 0
+    def hasNext = self.hasNext && i < n
+    def next() =
+      if (hasNext) {
+        i += 1
+        self.next()
+      }
+      else Iterator.empty.next()
+  }
+  def drop(n: Int): Iterator[A] = {
+    var i = 0
+    while (i < n && hasNext) {
+      next()
+      i += 1
+    }
+    this
+  }
+  def zip[B](that: IterableOnce[B]): Iterator[(A, B)] = new Iterator[(A, B)] {
+    val thatIterator = that.iterator()
+    def hasNext = self.hasNext && thatIterator.hasNext
+    def next() = (self.next(), thatIterator.next())
+  }
+}
+
+object Iterator {
+  val empty: Iterator[Nothing] = new Iterator[Nothing] {
+    def hasNext = false
+    def next() = throw new NoSuchElementException("next on empty iterator")
+  }
+  def apply[A](xs: A*): Iterator[A] = new IndexedView[A] {
+    val length = xs.length
+    def apply(n: Int) = xs(n)
+  }.iterator()
+}
